refactor: Repository 관련 클래스 - 인터페이스 분리하여 리팩토링
