+++ src/main/java/com/example/bcsd/repository/MemberRepositoryImpl.java
@@ -1,0 +1,57 @@
+package com.example.bcsd.repository;
+
+import com.example.bcsd.entity.Member;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public class MemberRepositoryImpl implements MemberRepository {
+
+    @PersistenceContext
+    private EntityManager em;
+
+    @Override
+    public List<Member> findAll() {
+        return em.createQuery("SELECT m FROM Member m", Member.class)
+                .getResultList();
+    }
+
+    @Override
+    public Optional<Member> findById(Long id) {
+        return Optional.ofNullable(em.find(Member.class, id));
+    }
+
+    @Override
+    public Member save(Member member) {
+        em.persist(member);
+        return member;
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        Member member = em.find(Member.class, id);
+        if (member != null) {
+            em.remove(member);
+        }
+    }
+
+    @Override
+    public int countArticlesByMember(Long memberId) {
+        Long count = em.createQuery("SELECT COUNT(a) FROM Article a WHERE a.writerId = :writerId", Long.class)
+                .setParameter("writerId", memberId)
+                .getSingleResult();
+        return count.intValue();
+    }
+
+    @Override
+    public boolean existsByEmail(String email) {
+        Long count = em.createQuery("SELECT COUNT(m) FROM Member m WHERE m.email = :email", Long.class)
+                .setParameter("email", email)
+                .getSingleResult();
+        return count > 0;
+    }
+}
